MESH-1037: Refactor identity pallet (#1144)
* identity: extract authorization

* simplify add_secondary_keys_with_authorization

* identity: nuke dead code

* extract base_invalidate_cdd_claims

* extract base_remove_secondary_keys

* extract base_add_secondary_keys_with_authorization

* move acc ref count -> keys

* move helpers -> keys

* move rpc stuff -> keys

* extract more -> keys

* identity: move more code

* identity: un-trait-ify functions

* identity: extract claims

* cargo fmt

* fix benchmarks

* identity/keys: optimize with ensure_primary_key
