+++ pallets/identity/src/lib.rs
@@ -79,7 +79,16 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 #![recursion_limit = "256"]
-#![feature(const_option, bool_to_option, option_result_contains)]
+#![feature(
+    const_option,
+    bool_to_option,
+    option_result_contains,
+    crate_visibility_modifier
+)]
+
+mod auth;
+mod claims;
+mod keys;
 
 pub mod types;
 pub use types::{
@@ -93,56 +102,34 @@ use codec::{Decode, Encode};
 use confidential_identity::ScopeClaimProof;
 use core::convert::From;
 use frame_support::{
-    debug, decl_error, decl_module, decl_storage,
-    dispatch::{DispatchError, DispatchResult},
-    ensure, fail,
+    decl_error, decl_module, decl_storage,
+    dispatch::DispatchResult,
     traits::{ChangeMembers, Currency, EnsureOrigin, Get, InitializeMembers},
     weights::{
         DispatchClass::{Normal, Operational},
         Pays,
     },
-    StorageDoubleMap,
 };
-use frame_system::{self as system, ensure_root, ensure_signed};
-use pallet_base::{ensure_length_ok, ensure_string_limited};
+use frame_system::ensure_root;
 pub use polymesh_common_utilities::traits::identity::WeightInfo;
 use polymesh_common_utilities::{
-    constants::did::{SECURITY_TOKEN, USER},
+    constants::did::SECURITY_TOKEN,
     protocol_fee::{ChargeProtocolFee, ProtocolOp},
-    traits::{
-        asset::AssetSubTrait,
-        group::{GroupTrait, InactiveMember},
-        identity::{
-            AuthorizationNonce, Config, IdentityFnTrait, RawEvent, SecondaryKeyWithAuth,
-            TargetIdAuthorization,
-        },
-        multisig::MultiSigSubTrait,
-        transaction_payment::CddAndFeeDetails,
-        AccountCallPermissionsData, CheckAccountCallPermissions,
+    traits::identity::{
+        AuthorizationNonce, Config, IdentityFnTrait, RawEvent, SecondaryKeyWithAuth,
     },
-    Context, SystematicIssuers, GC_DID, SYSTEMATIC_ISSUERS,
+    SystematicIssuers, GC_DID, SYSTEMATIC_ISSUERS,
 };
 use polymesh_primitives::{
-    extract_auth,
-    investor_zkproof_data::{v1::InvestorZKProofData, InvestorZKProofData as InvestorZKProof},
-    secondary_key::{self, api::LegacyPermissions},
-    storage_migration_ver, valid_proof_of_investor, Authorization, AuthorizationData,
-    AuthorizationError, AuthorizationType, CddId, Claim, ClaimType, DispatchableName,
-    ExtrinsicPermissions, Identity as DidRecord, IdentityClaim, IdentityId, InvestorUid,
-    PalletName, Permissions, Scope, ScopeId, SecondaryKey, Signatory, Ticker,
+    investor_zkproof_data::v1::InvestorZKProofData, secondary_key::api::LegacyPermissions,
+    storage_migration_ver, Authorization, AuthorizationData, AuthorizationType, CddId, Claim,
+    ClaimType, Identity as DidRecord, IdentityClaim, IdentityId, Permissions, Scope, SecondaryKey,
+    Signatory, Ticker,
 };
-use sp_core::sr25519::Signature;
-use sp_io::hashing::blake2_256;
-use sp_runtime::{
-    traits::{
-        AccountIdConversion, CheckedAdd, Hash, IdentifyAccount, SaturatedConversion, Verify, Zero,
-    },
-    AnySignature,
-};
-use sp_std::{convert::TryFrom, iter, mem::replace, prelude::*, vec};
+use sp_runtime::traits::Hash;
+use sp_std::{convert::TryFrom, iter, prelude::*, vec};
 
 pub type Event<T> = polymesh_common_utilities::traits::identity::Event<T>;
-type CallPermissions<T> = pallet_permissions::Module<T>;
 
 storage_migration_ver!(0);
 
@@ -284,13 +271,13 @@ decl_module! {
             target_account: T::AccountId,
             secondary_keys: Vec<SecondaryKey<T::AccountId>>
         ) {
-            let cdd_id = Self::ensure_perms(origin)?;
-            Self::base_cdd_register_did(cdd_id, target_account, secondary_keys)?;
+            Self::base_cdd_register_did(origin, target_account, secondary_keys)?;
         }
 
-        /// It invalidates any claim generated by `cdd` from `disable_from` timestamps.
-        /// You can also define an expiration time, which will invalidate all claims generated by
-        /// that `cdd` and remove it as CDD member group.
+        /// Invalidates any claim generated by `cdd` from `disable_from` timestamps.
+        ///
+        /// You can also define an expiration time,
+        /// which will invalidate all claims generated by that `cdd` and remove it as CDD member group.
         #[weight = (<T as Config>::WeightInfo::invalidate_cdd_claims(), Operational, Pays::Yes)]
         pub fn invalidate_cdd_claims(
             origin,
@@ -298,16 +285,7 @@ decl_module! {
             disable_from: T::Moment,
             expiry: Option<T::Moment>,
         ) {
-            ensure_root(origin)?;
-
-            let now = <pallet_timestamp::Module<T>>::get();
-            ensure!(
-                T::CddServiceProviders::get_valid_members_at(now).contains(&cdd),
-                Error::<T>::UnAuthorizedCddProvider
-            );
-
-            T::CddServiceProviders::disable_member(cdd, expiry, Some(disable_from))?;
-            Self::deposit_event(RawEvent::CddClaimsInvalidated(cdd, disable_from));
+            Self::base_invalidate_cdd_claims(origin, cdd, disable_from, expiry)?;
         }
 
         /// Removes specified secondary keys of a DID if present.
@@ -319,58 +297,7 @@ decl_module! {
         /// `950_000_000 + 60_000 * signers_to_remove.len()`
         #[weight = <T as Config>::WeightInfo::remove_secondary_keys(signers_to_remove.len() as u32)]
         pub fn remove_secondary_keys(origin, signers_to_remove: Vec<Signatory<T::AccountId>>) {
-            let PermissionedCallOriginData {
-                sender,
-                primary_did: did,
-                ..
-            } = Self::ensure_origin_call_permissions(origin)?;
-            let _grants_checked = Self::grant_check_only_primary_key(&sender, did)?;
-
-            // Ensure that it is safe to unlink the secondary keys from the did.
-            for signer in &signers_to_remove {
-                if let Signatory::Account(key) = &signer {
-                    Self::ensure_key_unlinkable_from_did(key)?;
-                }
-            }
-
-            // Remove links and get all authorization IDs per signer.
-            signers_to_remove
-                .iter()
-                .flat_map(|signer| {
-                    use either::Either::{Left, Right};
-
-                    // Unlink each of the given secondary keys from `did`.
-                    if let Signatory::Account(key) = &signer {
-                        // Unlink multisig signers.
-                        if T::MultiSig::is_multisig(key) {
-                            if !T::Balances::total_balance(key).is_zero() {
-                                return Left(iter::empty());
-                            }
-                            // Unlink multisig signers from the identity.
-                            Self::unlink_multisig_signers_from_did(
-                                T::MultiSig::get_key_signers(key),
-                                did
-                            );
-                        }
-                        // Unlink the secondary account key.
-                        Self::unlink_account_key_from_did(key, did);
-                    }
-
-                    // Compute list of `auth_id` from `did`.
-                    Right(
-                        <Authorizations<T>>::iter_prefix_values(signer)
-                            .filter_map(move |auth| (auth.authorized_by == did).then_some((signer, auth.auth_id)))
-                    )
-                })
-                // Remove authorizations.
-                .for_each(|(signer, auth_id)| Self::unsafe_remove_auth(signer, auth_id, &did, true));
-
-            // Update secondary keys at Identity.
-            <DidRecords<T>>::mutate(did, |record| {
-                (*record).remove_secondary_keys(&signers_to_remove);
-            });
-
-            Self::deposit_event(RawEvent::SecondaryKeysRemoved(did, signers_to_remove));
+            Self::base_remove_secondary_keys(origin, signers_to_remove)?;
         }
 
         /// Call this with the new primary key. By invoking this method, caller accepts authorization
@@ -442,24 +369,12 @@ decl_module! {
             Self::base_revoke_claim(target, claim_type, issuer, scope)
         }
 
-        /// It sets permissions for an specific `target_key` key.
+        /// Sets permissions for an specific `target_key` key.
+        ///
         /// Only the primary key of an identity is able to set secondary key permissions.
         #[weight = <T as Config>::WeightInfo::set_permission_to_signer()]
-        pub fn set_permission_to_signer(
-            origin,
-            signer: Signatory<T::AccountId>,
-            permissions: Permissions
-        ) -> DispatchResult {
-            let PermissionedCallOriginData {
-                sender,
-                primary_did: did,
-                ..
-            } = Self::ensure_origin_call_permissions(origin)?;
-            let record = Self::grant_check_only_primary_key(&sender, did)?;
-
-            // Ensure that the signer is a secondary key of the caller's Identity
-            ensure!(record.secondary_keys.iter().any(|si| si.signer == signer), Error::<T>::NotASigner);
-            Self::update_secondary_key_permissions(did, &signer, permissions)
+        pub fn set_permission_to_signer(origin, signer: Signatory<T::AccountId>, perms: Permissions) {
+            Self::base_set_permission_to_signer(origin, signer, perms)?;
         }
 
         /// This function is a workaround for https://github.com/polkadot-js/apps/issues/3632
@@ -495,14 +410,10 @@ decl_module! {
         pub fn add_authorization(
             origin,
             target: Signatory<T::AccountId>,
-            authorization_data: AuthorizationData<T::AccountId>,
+            data: AuthorizationData<T::AccountId>,
             expiry: Option<T::Moment>
         ) {
-            let from_did = Self::ensure_perms(origin)?;
-            if let AuthorizationData::JoinIdentity(perms) = &authorization_data {
-                Self::ensure_perms_length_limited(perms)?;
-            }
-            Self::add_auth(from_did, target, authorization_data, expiry);
+            Self::base_add_authorization(origin, target, data, expiry)?;
         }
 
         /// Removes an authorization.
@@ -514,21 +425,7 @@ decl_module! {
             auth_id: u64,
             _auth_issuer_pays: bool,
         ) {
-            let sender = ensure_signed(origin)?;
-            let from_did = if <KeyToIdentityIds<T>>::contains_key(&sender) {
-                // If the sender is linked to an identity, ensure that it has relevant permissions
-                CallPermissions::<T>::ensure_call_permissions(&sender)?.primary_did
-            } else {
-                Context::current_identity_or::<Self>(&sender)?
-            };
-
-            let auth = Self::ensure_authorization(&target, auth_id)?;
-            let revoked = auth.authorized_by == from_did;
-            ensure!(
-                revoked || target.eq_either(&from_did, &sender),
-                Error::<T>::Unauthorized
-            );
-            Self::unsafe_remove_auth(&target, auth_id, &auth.authorized_by, revoked);
+            Self::base_remove_authorization(origin, target, auth_id)?;
         }
 
         /// It adds secondary keys to target identity `id`.
@@ -549,75 +446,7 @@ decl_module! {
             additional_keys: Vec<SecondaryKeyWithAuth<T::AccountId>>,
             expires_at: T::Moment
         ) {
-            let PermissionedCallOriginData {
-                sender,
-                primary_did: did,
-                ..
-            } = Self::ensure_origin_call_permissions(origin)?;
-            let _grants_checked = Self::grant_check_only_primary_key(&sender, did)?;
-
-            // 0. Check expiration
-            let now = <pallet_timestamp::Module<T>>::get();
-            ensure!(now < expires_at, Error::<T>::AuthorizationExpired);
-            let authorization = TargetIdAuthorization {
-                target_id: did,
-                nonce: Self::offchain_authorization_nonce(did),
-                expires_at
-            };
-            let auth_encoded = authorization.encode();
-
-            let mut record = <DidRecords<T>>::get(did);
-
-            // Ensure we won't have too many keys.
-            ensure_length_ok::<T>(record.secondary_keys.len().saturating_add(additional_keys.len()))?;
-
-            // 1. Verify signatures.
-            for si_with_auth in additional_keys.iter() {
-                let si: SecondaryKey<T::AccountId> = si_with_auth.secondary_key.clone().into();
-
-                Self::ensure_perms_length_limited(&si.permissions)?;
-
-                // Get account_id from signer.
-                let account_id = si.signer.as_account().ok_or(Error::<T>::InvalidAccountKey)?;
-
-                // 1.1. Constraint 1-to-1 account to DID.
-                ensure!(Self::can_link_account_key_to_did(account_id), Error::<T>::AlreadyLinked);
-
-                // 1.2. Verify the signature.
-                let signature = AnySignature::from(Signature::from_h512(si_with_auth.auth_signature));
-                let signer: <<AnySignature as Verify>::Signer as IdentifyAccount>::AccountId =
-                    Decode::decode(&mut &account_id.encode()[..]).map_err(|_| {
-                        Error::<T>::CannotDecodeSignerAccountId
-                    })?;
-                ensure!(
-                    signature.verify(auth_encoded.as_slice(), &signer),
-                    Error::<T>::InvalidAuthorizationSignature
-                );
-            }
-            // 1.999. Charge the fee.
-            T::ProtocolFee::batch_charge_fee(
-                ProtocolOp::IdentityAddSecondaryKeysWithAuthorization,
-                additional_keys.len()
-            )?;
-            // 2.1. Link keys to identity
-            let additional_keys_si: Vec<_> =
-                additional_keys.into_iter()
-                .map(|si_with_auth| si_with_auth.secondary_key)
-                .collect();
-
-            additional_keys_si.iter().for_each(|sk| {
-                if let Signatory::Account(key) = &sk.signer {
-                    Self::link_account_key_to_did(key, did);
-                }
-            });
-            // 2.2. Update that identity information and its offchain authorization nonce.
-            record.add_secondary_keys(additional_keys_si.iter().map(|sk| sk.clone().into()));
-            <DidRecords<T>>::insert(did, record);
-            OffChainAuthorizationNonce::mutate(did, |offchain_nonce| {
-                *offchain_nonce = authorization.nonce + 1;
-            });
-
-            Self::deposit_event(RawEvent::SecondaryKeysAdded(did, additional_keys_si));
+            Self::base_add_secondary_keys_with_authorization(origin, additional_keys, expires_at)?;
         }
 
         /// Add `Claim::InvestorUniqueness` claim for a given target identity.
@@ -761,527 +590,6 @@ impl<T: Config> Module<T> {
         Self::deposit_event(RawEvent::AssetDidRegistered(did, ticker));
     }
 
-    pub fn ensure_no_id_record(id: IdentityId) -> DispatchResult {
-        ensure!(!Self::is_identity_exists(&id), Error::<T>::DidAlreadyExists);
-        Ok(())
-    }
-
-    /// Increment the reference counter for `key`.
-    pub fn add_account_key_ref_count(key: &T::AccountId) {
-        <AccountKeyRefCount<T>>::mutate(key, |n| *n = n.saturating_add(1_u64));
-    }
-
-    /// Decrement the reference counter for `key`.
-    pub fn remove_account_key_ref_count(key: &T::AccountId) {
-        <AccountKeyRefCount<T>>::mutate(key, |n| *n = n.saturating_sub(1_u64));
-    }
-
-    /// Ensure that the account key is safe to unlink from it's identity.
-    fn ensure_key_unlinkable_from_did(key: &T::AccountId) -> DispatchResult {
-        ensure!(
-            <AccountKeyRefCount<T>>::get(key) == 0,
-            Error::<T>::AccountKeyIsBeingUsed
-        );
-        Ok(())
-    }
-
-    /// Returns `Err(DidDoesNotExist)` unless `id` has an associated record.
-    fn ensure_id_record_exists(id: IdentityId) -> DispatchResult {
-        ensure!(Self::is_identity_exists(&id), Error::<T>::DidDoesNotExist);
-        Ok(())
-    }
-
-    /// Accepts an auth to join an identity as a signer
-    pub fn join_identity(origin: T::Origin, auth_id: u64) -> DispatchResult {
-        let key = ensure_signed(origin)?;
-        let signer = Signatory::Account(key.clone());
-        Self::accept_auth_with(&signer, auth_id, |data, target_did| {
-            let permissions = extract_auth!(data, JoinIdentity(p));
-            // Not really needed unless we allow identities to be deleted.
-            Self::ensure_id_record_exists(target_did)?;
-
-            // Link the secondary key.
-            Self::ensure_key_did_unlinked(&key)?;
-            // Check that the new Identity has a valid CDD claim.
-            ensure!(Self::has_valid_cdd(target_did), Error::<T>::TargetHasNoCdd);
-            // Charge the protocol fee after all checks.
-            T::ProtocolFee::charge_fee(ProtocolOp::IdentityAddSecondaryKeysWithAuthorization)?;
-            // Update current did of the transaction to the newly joined did.
-            // This comes handy when someone uses a batch transaction to leave their identity,
-            // join another identity, and then do something as the new identity.
-            T::CddHandler::set_current_identity(&target_did);
-
-            Self::unsafe_join_identity(target_did, permissions, key);
-            Ok(())
-        })
-    }
-
-    /// Ensure `key` isn't linked to a DID.
-    pub fn ensure_key_did_unlinked(key: &T::AccountId) -> DispatchResult {
-        ensure!(
-            Self::can_link_account_key_to_did(key),
-            Error::<T>::AlreadyLinked
-        );
-        Ok(())
-    }
-
-    /// Joins a DID as an account based secondary key.
-    pub fn unsafe_join_identity(
-        target_did: IdentityId,
-        permissions: Permissions,
-        key: T::AccountId,
-    ) {
-        Self::link_account_key_to_did(&key, target_did);
-
-        // Link the secondary key.
-        let sk = SecondaryKey::new(Signatory::Account(key), permissions);
-        <DidRecords<T>>::mutate(target_did, |identity| {
-            identity.add_secondary_keys(iter::once(sk.clone()));
-        });
-        Self::deposit_event(RawEvent::SecondaryKeysAdded(target_did, vec![sk.into()]));
-    }
-
-    /// Adds an authorization.
-    pub fn add_auth(
-        from: IdentityId,
-        target: Signatory<T::AccountId>,
-        authorization_data: AuthorizationData<T::AccountId>,
-        expiry: Option<T::Moment>,
-    ) -> u64 {
-        let new_nonce = Self::multi_purpose_nonce() + 1u64;
-        MultiPurposeNonce::put(&new_nonce);
-
-        let auth = Authorization {
-            authorization_data: authorization_data.clone(),
-            authorized_by: from,
-            expiry,
-            auth_id: new_nonce,
-        };
-
-        <Authorizations<T>>::insert(target.clone(), new_nonce, auth);
-        <AuthorizationsGiven<T>>::insert(from, new_nonce, target.clone());
-
-        // This event is split in order to help the event harvesters.
-        Self::deposit_event(RawEvent::AuthorizationAdded(
-            from,
-            target.as_identity().cloned(),
-            target.as_account().cloned(),
-            new_nonce,
-            authorization_data,
-            expiry,
-        ));
-
-        new_nonce
-    }
-
-    /// Removes any authorization. No questions asked.
-    /// NB: Please do all the required checks before calling this function.
-    fn unsafe_remove_auth(
-        target: &Signatory<T::AccountId>,
-        auth_id: u64,
-        authorizer: &IdentityId,
-        revoked: bool,
-    ) {
-        <Authorizations<T>>::remove(target, auth_id);
-        <AuthorizationsGiven<T>>::remove(authorizer, auth_id);
-        let id = target.as_identity().cloned();
-        let acc = target.as_account().cloned();
-        let event = if revoked {
-            RawEvent::AuthorizationRevoked
-        } else {
-            RawEvent::AuthorizationRejected
-        };
-        Self::deposit_event(event(id, acc, auth_id))
-    }
-
-    /// Given that `auth_by` is the DID that issued an authorization,
-    /// ensure that it matches `from`, or otherwise return an error.
-    pub fn ensure_auth_by(auth_by: IdentityId, from: IdentityId) -> DispatchResult {
-        ensure!(auth_by == from, AuthorizationError::Unauthorized);
-        Ok(())
-    }
-
-    /// Accepts an authorization `auth_id` as `target`,
-    /// executing `accepter` for case-specific additional validation and storage changes.
-    pub fn accept_auth_with(
-        target: &Signatory<T::AccountId>,
-        auth_id: u64,
-        accepter: impl FnOnce(AuthorizationData<T::AccountId>, IdentityId) -> DispatchResult,
-    ) -> DispatchResult {
-        // Extract authorization.
-        let auth = Self::ensure_authorization(target, auth_id)?;
-
-        // Ensure that `auth.expiry`, if provided, is in the future.
-        if let Some(expiry) = auth.expiry {
-            let now = <pallet_timestamp::Module<T>>::get();
-            ensure!(expiry > now, AuthorizationError::Expired);
-        }
-
-        // Run custom per-type validation and updates.
-        accepter(auth.authorization_data, auth.authorized_by)?;
-
-        // Remove authorization from storage and emit event.
-        <Authorizations<T>>::remove(&target, auth_id);
-        <AuthorizationsGiven<T>>::remove(auth.authorized_by, auth_id);
-        Self::deposit_event(RawEvent::AuthorizationConsumed(
-            target.as_identity().cloned(),
-            target.as_account().cloned(),
-            auth_id,
-        ));
-        Ok(())
-    }
-
-    /// Return and ensure that there's an authorization `auth_id` for `target`.
-    fn ensure_authorization(
-        target: &Signatory<T::AccountId>,
-        auth_id: u64,
-    ) -> Result<Authorization<T::AccountId, T::Moment>, DispatchError> {
-        Self::authorizations(target, auth_id).ok_or_else(|| AuthorizationError::Invalid.into())
-    }
-
-    /// Accepts a primary key rotation.
-    fn accept_primary_key_rotation(
-        origin: T::Origin,
-        rotation_auth_id: u64,
-        optional_cdd_auth_id: Option<u64>,
-    ) -> DispatchResult {
-        let sender = ensure_signed(origin)?;
-        let signer = Signatory::Account(sender.clone());
-        Self::accept_auth_with(&signer, rotation_auth_id, |data, target_did| {
-            // Ensure Authorization is a `RotatePrimaryKey`.
-            extract_auth!(data, RotatePrimaryKey);
-            Self::unsafe_primary_key_rotation(sender, target_did, optional_cdd_auth_id)
-        })
-    }
-
-    /// Processes primary key rotation.
-    pub fn unsafe_primary_key_rotation(
-        sender: T::AccountId,
-        rotation_for_did: IdentityId,
-        optional_cdd_auth_id: Option<u64>,
-    ) -> DispatchResult {
-        // Accept authorization from CDD service provider.
-        if Self::cdd_auth_for_primary_key_rotation() {
-            let auth_id = optional_cdd_auth_id
-                .ok_or_else(|| Error::<T>::InvalidAuthorizationFromCddProvider)?;
-
-            let signer = Signatory::Account(sender.clone());
-            Self::accept_auth_with(&signer, auth_id, |data, auth_by| {
-                let attestation_for_did = extract_auth!(data, AttestPrimaryKeyRotation(a));
-                // Attestor must be a CDD service provider.
-                ensure!(
-                    T::CddServiceProviders::is_member(&auth_by),
-                    Error::<T>::NotCddProviderAttestation
-                );
-                // Ensure authorizations are for the same DID.
-                ensure!(
-                    rotation_for_did == attestation_for_did,
-                    Error::<T>::AuthorizationsNotForSameDids
-                );
-                Ok(())
-            })?;
-        }
-
-        Self::ensure_key_did_unlinked(&sender)?;
-
-        // Get the current DidRecord.
-        let mut record = Self::did_records(&rotation_for_did);
-        let old_primary_key = record.primary_key;
-
-        // Ensure that it is safe to unlink the primary key from the did.
-        Self::ensure_key_unlinkable_from_did(&old_primary_key)?;
-
-        // Replace primary key of the owner that initiated key rotation.
-        Self::unlink_account_key_from_did(&old_primary_key, rotation_for_did);
-        record.primary_key = sender.clone();
-        Self::link_account_key_to_did(&sender, rotation_for_did);
-        <DidRecords<T>>::insert(&rotation_for_did, record);
-
-        Self::deposit_event(RawEvent::PrimaryKeyUpdated(
-            rotation_for_did,
-            old_primary_key,
-            sender,
-        ));
-        Ok(())
-    }
-
-    /// Updates permissions of secondary items.
-    /// Private and not sanitized function. It is designed to be used internally by
-    /// others sanitezed functions.
-    fn update_secondary_key_permissions(
-        target_did: IdentityId,
-        signer: &Signatory<T::AccountId>,
-        permissions: Permissions,
-    ) -> DispatchResult {
-        Self::ensure_perms_length_limited(&permissions)?;
-
-        <DidRecords<T>>::mutate(target_did, |record| {
-            if let Some(secondary_key) = (*record)
-                .secondary_keys
-                .iter_mut()
-                .find(|si| si.signer == *signer)
-            {
-                let old_perms = replace(&mut secondary_key.permissions, permissions.clone());
-                Self::deposit_event(RawEvent::SecondaryKeyPermissionsUpdated(
-                    target_did,
-                    secondary_key.clone().into(),
-                    old_perms,
-                    permissions,
-                ));
-            }
-        });
-
-        Ok(())
-    }
-
-    /// It checks if `key` is a secondary key of `did` identity.
-    /// # IMPORTANT
-    /// If secondary keys are frozen this function always returns false.
-    /// A primary key cannot be frozen.
-    pub fn is_key_authorized(did: IdentityId, key: &T::AccountId) -> bool {
-        let record = <DidRecords<T>>::get(did);
-
-        // Check primary id or key.
-        &record.primary_key == key
-            // Check secondary items if DID is not frozen.
-            || !Self::is_did_frozen(did) && record.secondary_keys.iter().any(|si| si.signer.as_account().contains(&key))
-    }
-
-    /// It checks if `key` is a secondary key of `did` identity.
-    pub fn is_signer(did: IdentityId, signer: &Signatory<T::AccountId>) -> bool {
-        let record = <DidRecords<T>>::get(did);
-        record.secondary_keys.iter().any(|si| si.signer == *signer)
-    }
-
-    /// Use `did` as reference.
-    pub fn is_primary_key(did: &IdentityId, key: &T::AccountId) -> bool {
-        key == &<DidRecords<T>>::get(did).primary_key
-    }
-
-    /// Ensure that any `Scope::Custom(data)` is limited to 32 characters.
-    pub fn ensure_custom_scopes_limited(claim: &Claim) -> DispatchResult {
-        if let Some(Scope::Custom(data)) = claim.as_scope() {
-            ensure!(data.len() <= 32, Error::<T>::CustomScopeTooLong);
-        }
-        Ok(())
-    }
-
-    /// It returns true if `id_claim` is not expired at `moment`.
-    #[inline]
-    fn is_identity_claim_not_expired_at(id_claim: &IdentityClaim, moment: T::Moment) -> bool {
-        if let Some(expiry) = id_claim.expiry {
-            expiry > moment.saturated_into::<u64>()
-        } else {
-            true
-        }
-    }
-
-    /// It fetches an specific `claim_type` claim type for target identity `id`, which was issued
-    /// by `issuer`.
-    /// It only returns non-expired claims.
-    pub fn fetch_claim(
-        id: IdentityId,
-        claim_type: ClaimType,
-        issuer: IdentityId,
-        scope: Option<Scope>,
-    ) -> Option<IdentityClaim> {
-        let now = <pallet_timestamp::Module<T>>::get();
-
-        Self::fetch_base_claim_with_issuer(id, claim_type, issuer, scope)
-            .into_iter()
-            .find(|c| Self::is_identity_claim_not_expired_at(c, now))
-    }
-
-    /// See `Self::fetch_cdd`.
-    #[inline]
-    pub fn has_valid_cdd(claim_for: IdentityId) -> bool {
-        // It will never happen in production but helpful during testing.
-        #[cfg(feature = "no_cdd")]
-        if T::CddServiceProviders::get_members().is_empty() {
-            return true;
-        }
-
-        Self::base_fetch_cdd(claim_for, T::Moment::zero(), None).is_some()
-    }
-
-    /// It returns the CDD identity which issued the current valid CDD claim for `claim_for`
-    /// identity.
-    /// # Parameters
-    /// * `leeway` : This leeway is added to now() before check if claim is expired.
-    ///
-    /// # Safety
-    ///
-    /// No state change is allowed in this function because this function is used within the RPC
-    /// calls.
-    pub fn fetch_cdd(claim_for: IdentityId, leeway: T::Moment) -> Option<IdentityId> {
-        Self::base_fetch_cdd(claim_for, leeway, None)
-    }
-
-    fn base_fetch_cdd(
-        claim_for: IdentityId,
-        leeway: T::Moment,
-        filter_cdd_id: Option<CddId>,
-    ) -> Option<IdentityId> {
-        Self::base_fetch_valid_cdd_claims(claim_for, leeway, filter_cdd_id)
-            .map(|id_claim| id_claim.claim_issuer)
-            .next()
-    }
-
-    pub fn base_fetch_valid_cdd_claims(
-        claim_for: IdentityId,
-        leeway: T::Moment,
-        filter_cdd_id: Option<CddId>,
-    ) -> impl Iterator<Item = IdentityClaim> {
-        let exp_with_leeway = <pallet_timestamp::Module<T>>::get()
-            .checked_add(&leeway)
-            .unwrap_or_default();
-
-        // Supressing `mut` warning since we need mut in `runtime-benchmarks` feature but not otherwise.
-        #[allow(unused_mut)]
-        let mut active_cdds_temp = T::CddServiceProviders::get_active_members();
-
-        // For the benchmarks, self cdd claims are allowed and hence the claim target is added to the cdd providers list.
-        #[cfg(feature = "runtime-benchmarks")]
-        active_cdds_temp.push(claim_for);
-
-        let active_cdds = active_cdds_temp;
-        let inactive_not_expired_cdds = T::CddServiceProviders::get_inactive_members()
-            .into_iter()
-            .filter(|cdd| !T::CddServiceProviders::is_member_expired(cdd, exp_with_leeway))
-            .collect::<Vec<_>>();
-
-        Self::fetch_base_claims(claim_for, ClaimType::CustomerDueDiligence).filter(
-            move |id_claim| {
-                if let Some(cdd_id) = &filter_cdd_id {
-                    if let Claim::CustomerDueDiligence(claim_cdd_id) = &id_claim.claim {
-                        if claim_cdd_id != cdd_id {
-                            return false;
-                        }
-                    }
-                }
-
-                Self::is_identity_cdd_claim_valid(
-                    id_claim,
-                    exp_with_leeway,
-                    &active_cdds,
-                    &inactive_not_expired_cdds,
-                )
-            },
-        )
-    }
-
-    /// A CDD claims is considered valid if:
-    /// * Claim is not expired at `exp_with_leeway` moment.
-    /// * Its issuer is valid, that means:
-    ///   * Issuer is an active CDD provider, or
-    ///   * Issuer is an inactive CDD provider but claim was updated/created before that it was
-    ///   deactivated.
-    fn is_identity_cdd_claim_valid(
-        id_claim: &IdentityClaim,
-        exp_with_leeway: T::Moment,
-        active_cdds: &[IdentityId],
-        inactive_not_expired_cdds: &[InactiveMember<T::Moment>],
-    ) -> bool {
-        Self::is_identity_claim_not_expired_at(id_claim, exp_with_leeway)
-            && (active_cdds.contains(&id_claim.claim_issuer)
-                || SYSTEMATIC_ISSUERS
-                    .iter()
-                    .any(|si| si.as_id() == id_claim.claim_issuer)
-                || inactive_not_expired_cdds
-                    .iter()
-                    .filter(|cdd| cdd.id == id_claim.claim_issuer)
-                    .any(|cdd| {
-                        id_claim.last_update_date < cdd.deactivated_at.saturated_into::<u64>()
-                    }))
-    }
-
-    /// It iterates over all claims of type `claim_type` for target `id` identity.
-    /// Please note that it could return expired claims.
-    fn fetch_base_claims<'a>(
-        target: IdentityId,
-        claim_type: ClaimType,
-    ) -> impl Iterator<Item = IdentityClaim> + 'a {
-        Claims::iter_prefix_values(Claim1stKey { target, claim_type })
-    }
-
-    /// It fetches an specific `claim_type` claim type for target identity `id`, which was issued
-    /// by `issuer`.
-    fn fetch_base_claim_with_issuer(
-        target: IdentityId,
-        claim_type: ClaimType,
-        issuer: IdentityId,
-        scope: Option<Scope>,
-    ) -> Option<IdentityClaim> {
-        let pk = Claim1stKey { target, claim_type };
-        let sk = Claim2ndKey { issuer, scope };
-        Claims::contains_key(&pk, &sk).then(|| Claims::get(&pk, &sk))
-    }
-
-    /// It checks that `sender` is the primary key of `did` Identifier and that
-    /// did exists.
-    /// # Return
-    /// A result object containing the `DidRecord` of `did`.
-    pub fn grant_check_only_primary_key(
-        sender: &T::AccountId,
-        did: IdentityId,
-    ) -> Result<DidRecord<T::AccountId>, DispatchError> {
-        Self::ensure_id_record_exists(did)?;
-        let record = <DidRecords<T>>::get(did);
-        ensure!(*sender == record.primary_key, Error::<T>::KeyNotAllowed);
-        Ok(record)
-    }
-
-    /// Returns the DID associated with `key`, if any,
-    /// assuming it is either the primary key or isn't frozen.
-    pub fn get_identity(key: &T::AccountId) -> Option<IdentityId> {
-        <KeyToIdentityIds<T>>::try_get(key)
-            .ok()
-            .filter(|did| !Self::is_did_frozen(did) || Self::is_primary_key(&did, key))
-    }
-
-    /// It freezes/unfreezes the target `did` identity.
-    ///
-    /// # Errors
-    /// Only primary key can freeze/unfreeze an identity.
-    fn set_frozen_secondary_key_flags(origin: T::Origin, freeze: bool) -> DispatchResult {
-        let data = Self::ensure_origin_call_permissions(origin)?;
-        let did = data.primary_did;
-        let _ = Self::grant_check_only_primary_key(&data.sender, did)?;
-
-        let event = if freeze {
-            IsDidFrozen::insert(&did, true);
-            RawEvent::SecondaryKeysFrozen
-        } else {
-            IsDidFrozen::remove(&did);
-            RawEvent::SecondaryKeysUnfrozen
-        };
-        Self::deposit_event(event(did));
-        Ok(())
-    }
-
-    /// Checks that a key is not linked to any identity or multisig.
-    pub fn can_link_account_key_to_did(key: &T::AccountId) -> bool {
-        !<KeyToIdentityIds<T>>::contains_key(key) && !T::MultiSig::is_signer(key)
-    }
-
-    /// Links a primary or secondary `AccountId` key `key` to an identity `did`.
-    ///
-    /// This function applies the change if `can_link_account_key_to_did` returns `true`.
-    /// Otherwise, it does nothing.
-    pub fn link_account_key_to_did(key: &T::AccountId, did: IdentityId) {
-        if !<KeyToIdentityIds<T>>::contains_key(key) {
-            // `key` is not yet linked to any identity, so no constraints.
-            <KeyToIdentityIds<T>>::insert(key, did);
-        }
-    }
-
-    /// Unlinks an `AccountId` key `key` from an identity `did`.
-    fn unlink_account_key_from_did(key: &T::AccountId, did: IdentityId) {
-        if <KeyToIdentityIds<T>>::contains_key(key) && <KeyToIdentityIds<T>>::get(key) == did {
-            <KeyToIdentityIds<T>>::remove(key)
-        }
-    }
-
     /// IMPORTANT: No state change is allowed in this function
     /// because this function is used within the RPC calls
     /// It is a helper function that can be used to get did for any asset
@@ -1289,477 +597,6 @@ impl<T: Config> Module<T> {
         let mut buf = SECURITY_TOKEN.encode();
         buf.append(&mut ticker.encode());
         IdentityId::try_from(T::Hashing::hash(&buf[..]).as_ref())
-    }
-
-    /// Registers a did without adding a CDD claim for it.
-    pub fn _register_did(
-        sender: T::AccountId,
-        secondary_keys: Vec<SecondaryKey<T::AccountId>>,
-        protocol_fee_data: Option<ProtocolOp>,
-    ) -> Result<IdentityId, DispatchError> {
-        // Adding extrensic count to did nonce for some unpredictability
-        // NB: this does not guarantee randomness
-        let new_nonce =
-            Self::multi_purpose_nonce() + u64::from(<system::Module<T>>::extrinsic_count()) + 7u64;
-        // Even if this transaction fails, nonce should be increased for added unpredictability of dids
-        MultiPurposeNonce::put(&new_nonce);
-
-        // 1 Check constraints.
-        // Primary key is not linked to any identity.
-        Self::ensure_key_did_unlinked(&sender)?;
-        // Primary key is not part of secondary keys.
-        ensure!(
-            !secondary_keys
-                .iter()
-                .any(|sk| sk.signer.as_account() == Some(&sender)),
-            Error::<T>::SecondaryKeysContainPrimaryKey
-        );
-
-        let block_hash = <system::Module<T>>::block_hash(<system::Module<T>>::block_number());
-        let did = IdentityId(blake2_256(&(USER, block_hash, new_nonce).encode()));
-
-        // Make sure there's no pre-existing entry for the DID
-        // This should never happen but just being defensive here
-        Self::ensure_no_id_record(did)?;
-
-        // Secondary keys can be linked to the new identity.
-        for sk in &secondary_keys {
-            if let Signatory::Account(ref key) = sk.signer {
-                Self::ensure_key_did_unlinked(key)?;
-            }
-        }
-
-        // Charge the given fee.
-        if let Some(op) = protocol_fee_data {
-            T::ProtocolFee::charge_fee(op)?;
-        }
-
-        // 2. Apply changes to our extrinsic.
-        // 2.1. Link primary key and add pre-authorized secondary keys.
-        Self::link_account_key_to_did(&sender, did);
-        secondary_keys.iter().for_each(|sk| {
-            let data = AuthorizationData::JoinIdentity(sk.permissions.clone().into());
-            Self::add_auth(did, sk.signer.clone(), data, None);
-        });
-
-        // 2.2. Create a new identity record.
-        let record = DidRecord {
-            primary_key: sender.clone(),
-            ..Default::default()
-        };
-        <DidRecords<T>>::insert(&did, record);
-
-        // 2.3. Give `InitialPOLYX` to the primary key for testing.
-        T::Balances::deposit_creating(&sender, T::InitialPOLYX::get().into());
-
-        Self::deposit_event(RawEvent::DidCreated(
-            did,
-            sender,
-            secondary_keys
-                .into_iter()
-                .map(secondary_key::api::SecondaryKey::from)
-                .collect(),
-        ));
-        Ok(did)
-    }
-
-    /// Ensures length limits are enforced in `perms`.
-    fn ensure_perms_length_limited(perms: &Permissions) -> DispatchResult {
-        ensure_length_ok::<T>(perms.asset.complexity())?;
-        ensure_length_ok::<T>(perms.portfolio.complexity())?;
-        Self::ensure_extrinsic_perms_length_limited(&perms.extrinsic)
-    }
-
-    /// Ensures length limits are enforced in `perms`.
-    pub fn ensure_extrinsic_perms_length_limited(perms: &ExtrinsicPermissions) -> DispatchResult {
-        if let Some(set) = perms.inner() {
-            ensure_length_ok::<T>(set.len())?;
-            for elem in set {
-                ensure_string_limited::<T>(&elem.pallet_name)?;
-                if let Some(set) = elem.dispatchable_names.inner() {
-                    ensure_length_ok::<T>(set.len())?;
-                    for elem in set {
-                        ensure_string_limited::<T>(elem)?;
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// It adds a new claim without any previous security check.
-    pub fn base_add_claim(
-        target: IdentityId,
-        claim: Claim,
-        issuer: IdentityId,
-        expiry: Option<T::Moment>,
-    ) {
-        let inner_scope = claim.as_scope().cloned();
-        Self::base_add_claim_with_scope(target, claim, inner_scope, issuer, expiry)
-    }
-
-    /// Adds claims with no inner scope.
-    fn base_add_claim_with_scope(
-        target: IdentityId,
-        claim: Claim,
-        scope: Option<Scope>,
-        issuer: IdentityId,
-        expiry: Option<T::Moment>,
-    ) {
-        let claim_type = claim.claim_type();
-        let last_update_date = <pallet_timestamp::Module<T>>::get().saturated_into::<u64>();
-        let issuance_date = Self::fetch_claim(target, claim_type, issuer, scope.clone())
-            .map_or(last_update_date, |id_claim| id_claim.issuance_date);
-
-        let expiry = expiry.map(|m| m.saturated_into::<u64>());
-        let (pk, sk) = Self::get_claim_keys(target, claim_type, issuer, scope);
-        let id_claim = IdentityClaim {
-            claim_issuer: issuer,
-            issuance_date,
-            last_update_date,
-            expiry,
-            claim,
-        };
-
-        Claims::insert(&pk, &sk, id_claim.clone());
-        Self::deposit_event(RawEvent::ClaimAdded(target, id_claim));
-    }
-
-    /// Returns claim keys.
-    pub fn get_claim_keys(
-        target: IdentityId,
-        claim_type: ClaimType,
-        issuer: IdentityId,
-        scope: Option<Scope>,
-    ) -> (Claim1stKey, Claim2ndKey) {
-        let pk = Claim1stKey { target, claim_type };
-        let sk = Claim2ndKey { issuer, scope };
-        (pk, sk)
-    }
-
-    /// It ensures that CDD claim issuer is a valid CDD provider before add the claim.
-    ///
-    /// # Errors
-    /// - 'UnAuthorizedCddProvider' is returned if `issuer` is not a CDD provider.
-    fn base_add_cdd_claim(
-        target: IdentityId,
-        claim: Claim,
-        issuer: IdentityId,
-        expiry: Option<T::Moment>,
-    ) -> DispatchResult {
-        Self::ensure_authorized_cdd_provider(issuer)?;
-        // Ensure cdd_id uniqueness for a given target DID.
-        Self::ensure_cdd_id_validness(&claim, target)?;
-
-        Self::base_add_claim(target, claim, issuer, expiry);
-        Ok(())
-    }
-
-    /// Enforce CDD_ID uniqueness for a given target DID.
-    ///
-    /// # Errors
-    /// - `CDDIdNotUniqueForIdentity` is returned when new cdd claim's cdd_id doesn't match the existing cdd claim's cdd_id.
-    fn ensure_cdd_id_validness(claim: &Claim, target: IdentityId) -> DispatchResult {
-        if let Claim::CustomerDueDiligence(cdd_id) = claim {
-            ensure!(
-                cdd_id.is_default_cdd()
-                    || Self::base_fetch_valid_cdd_claims(target, 0u32.into(), None)
-                        .filter_map(|c| match c.claim {
-                            Claim::CustomerDueDiligence(c_id) => Some(c_id),
-                            _ => None,
-                        })
-                        .all(|c_id| c_id.is_default_cdd() || c_id == *cdd_id),
-                Error::<T>::CDDIdNotUniqueForIdentity
-            );
-        }
-        Ok(())
-    }
-
-    /// Decodes needed fields from `claim` and `proof`, and ensures that both are in the same
-    /// version.
-    ///
-    /// # Errors
-    /// - `ClaimVariantNotAllowed` if `claim` is not a `Claim::InvestorUniqueness` neither
-    /// `Claim::InvestorUniquenessV2`.
-    /// - `ClaimAndProofVersionsDoNotMatch` if `claim` and `proof` are different versions.
-    fn decode_investor_uniqueness_claim<'a>(
-        claim: &'a Claim,
-        proof: &'_ InvestorZKProof,
-        scope: Option<&'a Scope>,
-    ) -> Result<(&'a Scope, ScopeId, &'a CddId), DispatchError> {
-        let decode = match &claim {
-            Claim::InvestorUniqueness(scope, scope_id, cdd_id) => {
-                ensure!(
-                    matches!(proof, InvestorZKProof::V1(..)),
-                    Error::<T>::ClaimAndProofVersionsDoNotMatch
-                );
-                (scope, scope_id.clone(), cdd_id)
-            }
-            Claim::InvestorUniquenessV2(cdd_id) => match proof {
-                InvestorZKProof::V2(inner_proof) => (
-                    scope.ok_or(Error::<T>::InvalidScopeClaim)?,
-                    inner_proof.0.scope_id.compress().to_bytes().into(),
-                    cdd_id,
-                ),
-                _ => fail!(Error::<T>::ClaimAndProofVersionsDoNotMatch),
-            },
-            _ => fail!(Error::<T>::ClaimVariantNotAllowed),
-        };
-
-        Ok(decode)
-    }
-
-    /// # Errors
-    /// - 'ConfidentialScopeClaimNotAllowed` if :
-    ///     - Sender is not the issuer. That claim can be only added by your-self.
-    ///     - You are not the owner of that CDD_ID.
-    ///     - If claim is not valid.
-    ///
-    fn base_add_investor_uniqueness_claim(
-        origin: T::Origin,
-        target: IdentityId,
-        claim: Claim,
-        scope_opt: Option<Scope>,
-        proof: InvestorZKProof,
-        expiry: Option<T::Moment>,
-    ) -> DispatchResult {
-        Self::ensure_custom_scopes_limited(&claim)?;
-
-        // Decode needed fields and ensures `claim` is `InvestorUniqueness*`.
-        let (scope, scope_id, cdd_id) =
-            Self::decode_investor_uniqueness_claim(&claim, &proof, scope_opt.as_ref())?;
-
-        // Only owner of the identity can add that confidential claim.
-        let issuer = Self::ensure_signed_and_validate_claim_target(origin, target)?;
-        ensure!(
-            issuer == target,
-            Error::<T>::ConfidentialScopeClaimNotAllowed
-        );
-        // Verify the owner of that CDD_ID.
-        ensure!(
-            Self::base_fetch_cdd(target, T::Moment::zero(), Some(*cdd_id)).is_some(),
-            Error::<T>::ConfidentialScopeClaimNotAllowed
-        );
-
-        // Verify the confidential claim.
-        ensure!(
-            valid_proof_of_investor::evaluate_claim(scope, &claim, &target, &proof),
-            Error::<T>::InvalidScopeClaim
-        );
-
-        if let Scope::Ticker(ticker) = scope {
-            // Update the balance of the IdentityId under the ScopeId provided in claim data.
-            T::AssetSubTraitTarget::update_balance_of_scope_id(scope_id, target, *ticker);
-        }
-
-        let scope = Some(scope.clone());
-        Self::base_add_claim_with_scope(target, claim, scope, issuer, expiry);
-        Ok(())
-    }
-
-    /// Does the identity given by `did` exist?
-    pub fn is_identity_exists(did: &IdentityId) -> bool {
-        <DidRecords<T>>::contains_key(did)
-    }
-
-    /// Returns the record corresponding to `id`, if it exists.
-    fn identity_record_of(did: IdentityId) -> Option<DidRecord<T::AccountId>> {
-        Self::is_identity_exists(&did).then(|| <DidRecords<T>>::get(did))
-    }
-
-    /// It removes a claim from `target` which was issued by `issuer` without any security check.
-    fn base_revoke_claim(
-        target: IdentityId,
-        claim_type: ClaimType,
-        issuer: IdentityId,
-        scope: Option<Scope>,
-    ) -> DispatchResult {
-        let (pk, sk) = Self::get_claim_keys(target, claim_type, issuer, scope);
-
-        let investor_unique_scope_id = match Claims::get(&pk, &sk).claim {
-            Claim::InvestorUniqueness(_, scope_id, _) => Some(scope_id),
-            Claim::InvestorUniquenessV2(..) => match &sk.scope {
-                Some(Scope::Ticker(ticker)) => {
-                    Some(T::AssetSubTraitTarget::scope_id(ticker, &target))
-                }
-                _ => None,
-            },
-            _ => None,
-        };
-
-        // Only if claim is a `InvestorUniqueness*`.
-        if let Some(scope_id) = investor_unique_scope_id {
-            // Ensure the target is the issuer of the claim.
-            ensure!(
-                target == issuer,
-                Error::<T>::ConfidentialScopeClaimNotAllowed
-            );
-
-            // Ensure that the target has balance at scope = 0.
-            ensure!(
-                T::AssetSubTraitTarget::balance_of_at_scope(&scope_id, &target) == Zero::zero(),
-                Error::<T>::TargetHasNonZeroBalanceAtScopeId
-            );
-        }
-
-        let claim = Claims::take(&pk, &sk);
-        Self::deposit_event(RawEvent::ClaimRevoked(target, claim));
-        Ok(())
-    }
-
-    /// Returns an auth id if it is present and not expired.
-    pub fn get_non_expired_auth(
-        target: &Signatory<T::AccountId>,
-        auth_id: &u64,
-    ) -> Option<Authorization<T::AccountId, T::Moment>> {
-        Self::authorizations(target, *auth_id).filter(|auth| {
-            auth.expiry
-                .filter(|&expiry| <pallet_timestamp::Module<T>>::get() > expiry)
-                .is_none()
-        })
-    }
-
-    fn leave_identity(origin: T::Origin) -> DispatchResult {
-        let (key, did) = Self::ensure_did(origin)?;
-        let signer = Signatory::Account(key.clone());
-        ensure!(Self::is_signer(did, &signer), Error::<T>::NotASigner);
-
-        // Ensure that it is safe to unlink the account key from the did.
-        Self::ensure_key_unlinkable_from_did(&key)?;
-
-        // Unlink multisig signers.
-        if T::MultiSig::is_multisig(&key) {
-            ensure!(
-                T::Balances::total_balance(&key).is_zero(),
-                Error::<T>::MultiSigHasBalance
-            );
-            // Unlink multisig signers from the identity.
-            Self::unlink_multisig_signers_from_did(T::MultiSig::get_key_signers(&key), did);
-        }
-        Self::unlink_account_key_from_did(&key, did);
-
-        // Update secondary keys at Identity.
-        <DidRecords<T>>::mutate(did, |record| {
-            record.remove_secondary_keys(&[signer.clone()]);
-        });
-        Self::deposit_event(RawEvent::SignerLeft(did, signer));
-        Ok(())
-    }
-
-    fn unlink_multisig_signers_from_did(signers: Vec<T::AccountId>, did: IdentityId) {
-        for signer in signers {
-            Self::unlink_account_key_from_did(&signer, did)
-        }
-    }
-
-    /// Ensure that the origin is signed and that the given `target` is already in the system.
-    fn ensure_signed_and_validate_claim_target(
-        origin: T::Origin,
-        target: IdentityId,
-    ) -> Result<IdentityId, DispatchError> {
-        let primary_did = Self::ensure_perms(origin)?;
-        ensure!(
-            <DidRecords<T>>::contains_key(target),
-            Error::<T>::DidMustAlreadyExist
-        );
-        Ok(primary_did)
-    }
-
-    /// Checks whether the sender and the receiver of a transfer have valid investor uniqueness claims for a given ticker
-    pub fn verify_iu_claims_for_transfer(
-        ticker: Ticker,
-        from_did: IdentityId,
-        to_did: IdentityId,
-    ) -> bool {
-        let asset_scope = Some(Scope::from(ticker));
-        Self::base_verify_iu_claim(asset_scope.clone(), from_did)
-            && Self::base_verify_iu_claim(asset_scope, to_did)
-    }
-
-    /// Checks whether the identity has a valid investor uniqueness claim for a given ticker
-    pub fn verify_iu_claim(ticker: Ticker, did: IdentityId) -> bool {
-        let asset_scope = Some(Scope::from(ticker));
-        Self::base_verify_iu_claim(asset_scope, did)
-    }
-
-    fn base_verify_iu_claim(scope: Option<Scope>, did: IdentityId) -> bool {
-        Self::fetch_claim(did, ClaimType::InvestorUniqueness, did, scope.clone()).is_some()
-            || Self::fetch_claim(did, ClaimType::InvestorUniquenessV2, did, scope).is_some()
-    }
-}
-
-impl<T: Config> Module<T> {
-    /// RPC call to fetch some aggregate account data for fewer round trips.
-    pub fn get_key_identity_data(acc: T::AccountId) -> Option<types::KeyIdentityData<IdentityId>> {
-        let identity = Self::get_identity(&acc)?;
-        let record = <DidRecords<T>>::get(identity);
-        let permissions = if acc == record.primary_key {
-            None
-        } else {
-            Some(record.secondary_keys.into_iter().find_map(|sk| {
-                sk.signer.as_account().filter(|&a| a == &acc)?;
-                Some(sk.permissions)
-            })?)
-        };
-        Some(types::KeyIdentityData {
-            identity,
-            permissions,
-        })
-    }
-
-    /// RPC call to know whether the given did has valid cdd claim or not
-    pub fn is_identity_has_valid_cdd(
-        target: IdentityId,
-        leeway: Option<T::Moment>,
-    ) -> Option<IdentityId> {
-        Self::fetch_cdd(target, leeway.unwrap_or_default())
-    }
-
-    /// RPC call to query the given ticker did
-    pub fn get_asset_did(ticker: Ticker) -> Result<IdentityId, &'static str> {
-        Self::get_token_did(&ticker)
-    }
-
-    /// Retrieve DidRecords for `did`
-    pub fn get_did_records(
-        did: IdentityId,
-    ) -> RpcDidRecords<T::AccountId, SecondaryKey<T::AccountId>> {
-        if let Some(record) = Self::identity_record_of(did) {
-            RpcDidRecords::Success {
-                primary_key: record.primary_key,
-                secondary_keys: record.secondary_keys,
-            }
-        } else {
-            RpcDidRecords::IdNotFound
-        }
-    }
-
-    /// Use to get the filtered authorization data for a given signatory
-    /// - if auth_type is None then return authorizations data on the basis of the `allow_expired` boolean
-    /// - if auth_type is Some(value) then return filtered authorizations on the value basis type in conjunction
-    ///   with `allow_expired` boolean condition
-    pub fn get_filtered_authorizations(
-        signatory: Signatory<T::AccountId>,
-        allow_expired: bool,
-        auth_type: Option<AuthorizationType>,
-    ) -> Vec<Authorization<T::AccountId, T::Moment>> {
-        let now = <pallet_timestamp::Module<T>>::get();
-        let auths = <Authorizations<T>>::iter_prefix_values(signatory)
-            .filter(|auth| allow_expired || auth.expiry.filter(|&e| e < now).is_none());
-        if let Some(auth_type) = auth_type {
-            auths
-                .filter(|auth| Self::get_type(auth.authorization_data.clone(), auth_type.clone()))
-                .collect()
-        } else {
-            auths.collect()
-        }
-    }
-
-    pub fn get_type(
-        authorization_data: AuthorizationData<T::AccountId>,
-        type_of_auth: AuthorizationType,
-    ) -> bool {
-        type_of_auth == authorization_data.auth_type()
     }
 
     pub fn get_did_status(dids: Vec<IdentityId>) -> Vec<DidStatus> {
@@ -1770,117 +607,13 @@ impl<T: Config> Module<T> {
                     DidStatus::Unknown
                 }
                 // DID exists, but does it have a valid CDD?
-                else if Self::fetch_cdd(did, T::Moment::zero()).is_some() {
+                else if Self::has_valid_cdd(did) {
                     DidStatus::CddVerified
                 } else {
                     DidStatus::Exists
                 }
             })
             .collect()
-    }
-
-    /// Registers the systematic issuer with its DID.
-    #[allow(dead_code)]
-    fn register_systematic_id(issuer: SystematicIssuers)
-    where
-        <T as frame_system::Config>::AccountId: core::fmt::Display,
-    {
-        let acc = issuer.as_module_id().into_account();
-        let id = issuer.as_id();
-        debug::info!(
-            "Register Systematic id {} with account {} as {}",
-            issuer,
-            acc,
-            id
-        );
-
-        Self::do_register_id(acc, id, vec![]);
-    }
-
-    /// Registers `primary_key` as `id` identity.
-    #[allow(dead_code)]
-    fn do_register_id(
-        primary_key: T::AccountId,
-        id: IdentityId,
-        secondary_keys: Vec<SecondaryKey<T::AccountId>>,
-    ) {
-        <Module<T>>::link_account_key_to_did(&primary_key, id);
-        for sk in &secondary_keys {
-            if let Signatory::Account(key) = &sk.signer {
-                Self::link_account_key_to_did(key, id);
-            }
-        }
-
-        let record = DidRecord {
-            primary_key: primary_key.clone(),
-            secondary_keys,
-            ..Default::default()
-        };
-        <DidRecords<T>>::insert(&id, record);
-
-        Self::deposit_event(RawEvent::DidCreated(id, primary_key, vec![]));
-    }
-
-    /// Ensures that `origin`'s key is linked to a DID and returns both.
-    pub fn ensure_did(origin: T::Origin) -> Result<(T::AccountId, IdentityId), DispatchError> {
-        let sender = ensure_signed(origin)?;
-        let did = Context::current_identity_or::<Self>(&sender)?;
-        Ok((sender, did))
-    }
-
-    /// Checks call permissions and, if successful, returns the caller's account, primary and secondary identities.
-    pub fn ensure_origin_call_permissions(
-        origin: <T as frame_system::Config>::Origin,
-    ) -> Result<PermissionedCallOriginData<T::AccountId>, DispatchError> {
-        let sender = ensure_signed(origin)?;
-        let AccountCallPermissionsData {
-            primary_did,
-            secondary_key,
-        } = CallPermissions::<T>::ensure_call_permissions(&sender)?;
-        Ok(PermissionedCallOriginData {
-            sender,
-            primary_did,
-            secondary_key,
-        })
-    }
-
-    /// Ensure `origin` is signed and permissioned for this call, returning its DID.
-    pub fn ensure_perms(origin: T::Origin) -> Result<IdentityId, DispatchError> {
-        Self::ensure_origin_call_permissions(origin).map(|x| x.primary_did)
-    }
-
-    /// Ensures that the did is an active CDD Provider.
-    fn ensure_authorized_cdd_provider(did: IdentityId) -> DispatchResult {
-        ensure!(
-            T::CddServiceProviders::get_members().contains(&did),
-            Error::<T>::UnAuthorizedCddProvider
-        );
-        Ok(())
-    }
-
-    /// Ensures that the caller is an active CDD provider and creates a new did for the target.
-    /// This function returns the new did of the target.
-    ///
-    /// # Failure
-    /// - `origin` has to be a active CDD provider. Inactive CDD providers cannot add new
-    /// claims.
-    /// - `target_account` (primary key of the new Identity) can be linked to just one and only
-    /// one identity.
-    /// - External secondary keys can be linked to just one identity.
-    pub fn base_cdd_register_did(
-        caller_did: IdentityId,
-        target_account: T::AccountId,
-        secondary_keys: Vec<SecondaryKey<T::AccountId>>,
-    ) -> Result<IdentityId, DispatchError> {
-        // Sender has to be part of CDDProviders
-        Self::ensure_authorized_cdd_provider(caller_did)?;
-
-        // Register Identity
-        Self::_register_did(
-            target_account,
-            secondary_keys,
-            Some(ProtocolOp::IdentityCddRegisterDid),
-        )
     }
 
     #[cfg(feature = "runtime-benchmarks")]
@@ -1897,7 +630,7 @@ impl<T: Config> Module<T> {
     #[cfg(feature = "runtime-benchmarks")]
     /// Sets the current did in the context
     pub fn set_context_did(did: Option<IdentityId>) {
-        Context::set_current_identity::<Self>(did);
+        polymesh_common_utilities::Context::set_current_identity::<Self>(did);
     }
 }
 
@@ -1935,37 +668,6 @@ impl<T: Config> IdentityFnTrait<T::AccountId> for Module<T> {
         }
     }
 
-    /// Checks if the signer is authorized.
-    fn is_key_authorized(did: IdentityId, key: &T::AccountId) -> bool {
-        Self::is_key_authorized(did, key)
-    }
-
-    /// Checks if the keys is the primary key of the identity.
-    fn is_primary_key(did: &IdentityId, key: &T::AccountId) -> bool {
-        Self::is_primary_key(did, key)
-    }
-
-    /// Adds systematic CDD claims.
-    fn add_systematic_cdd_claims(targets: &[IdentityId], issuer: SystematicIssuers) {
-        for new_member in targets {
-            let cdd_id = CddId::new_v1(new_member.clone(), InvestorUid::from(new_member.as_ref()));
-            let cdd_claim = Claim::CustomerDueDiligence(cdd_id);
-            Self::base_add_claim(*new_member, cdd_claim, issuer.as_id(), None);
-        }
-    }
-
-    /// Removes systematic CDD claims.
-    fn revoke_systematic_cdd_claims(targets: &[IdentityId], issuer: SystematicIssuers) {
-        targets.iter().for_each(|removed_member| {
-            let _ = Self::base_revoke_claim(
-                *removed_member,
-                ClaimType::CustomerDueDiligence,
-                issuer.as_id(),
-                None,
-            );
-        });
-    }
-
     /// Provides the DID status for the given DID
     fn has_valid_cdd(target_did: IdentityId) -> bool {
         Self::has_valid_cdd(target_did)
@@ -1990,44 +692,6 @@ impl<T: Config> InitializeMembers<IdentityId> for Module<T> {
     /// Initializes members of a group by adding systematic claims for them.
     fn initialize_members(members: &[IdentityId]) {
         Self::add_systematic_cdd_claims(members, SystematicIssuers::CDDProvider);
-    }
-}
-
-impl<T: Config> CheckAccountCallPermissions<T::AccountId> for Module<T> {
-    // For weighting purposes, the function reads 4 storage values.
-    fn check_account_call_permissions(
-        who: &T::AccountId,
-        pallet_name: &PalletName,
-        function_name: &DispatchableName,
-    ) -> Option<AccountCallPermissionsData<T::AccountId>> {
-        // `who` is the original origin (signer) of the original extrinsic.
-        let primary_did = <KeyToIdentityIds<T>>::try_get(who).ok()?;
-        // The DID record must be present.
-        let primary_did_record = <DidRecords<T>>::try_get(&primary_did).ok()?;
-
-        if who == &primary_did_record.primary_key {
-            // It is a direct call and `who` is the primary key.
-            return Some(AccountCallPermissionsData {
-                primary_did,
-                secondary_key: None,
-            });
-        }
-
-        // DIDs with frozen secondary keys, AKA frozen DIDs, are not permitted to call extrinsics.
-        if Self::is_did_frozen(&primary_did) {
-            return None;
-        }
-
-        // Find the secondary key matching `who` and ensure it has sufficient permissions.
-        primary_did_record
-            .secondary_keys
-            .into_iter()
-            .find(|sk| sk.signer.as_account().contains(&who))
-            .filter(|sk| sk.has_extrinsic_permission(pallet_name, function_name))
-            .map(|sk| AccountCallPermissionsData {
-                primary_did,
-                secondary_key: Some(sk),
-            })
     }
 }
