+++ pallets/test-utils/src/lib.rs
@@ -146,13 +146,12 @@ decl_module! {
         /// one identity.
         #[weight = <T as Config>::WeightInfo::mock_cdd_register_did()]
         pub fn mock_cdd_register_did(origin, target_account: T::AccountId) {
-            let cdd_id = Identity::<T>::ensure_perms(origin)?;
-            let target_did = Identity::<T>::base_cdd_register_did(cdd_id, target_account, vec![])?;
+            let (cdd_did, target_did) = Identity::<T>::base_cdd_register_did(origin, target_account, vec![])?;
             let target_uid = confidential_identity::mocked::make_investor_uid(target_did.as_bytes());
 
             // Add CDD claim for the target
             let cdd_claim = Claim::CustomerDueDiligence(CddId::new_v1(target_did, target_uid.clone().into()));
-            Identity::<T>::base_add_claim(target_did, cdd_claim, cdd_id, None);
+            Identity::<T>::base_add_claim(target_did, cdd_claim, cdd_did, None);
 
             Self::deposit_event(RawEvent::MockInvestorUIDCreated(target_did, target_uid.into()));
         }
